requirements.txt
pendulum==1.3.2
pytest==3.2.5
pytest-mock==1.6.3
requests==2.18.4
